+++ opentelemetry-api/src/opentelemetry/context/contextvars_context.py
@@ -11,7 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from contextvars import ContextVar
+
+from __future__ import annotations
+
+from contextvars import ContextVar, Token
 
 from opentelemetry.context.context import Context, _RuntimeContext
 
@@ -28,7 +31,7 @@ class ContextVarsRuntimeContext(_RuntimeContext):
             self._CONTEXT_KEY, default=Context()
         )
 
-    def attach(self, context: Context) -> object:
+    def attach(self, context: Context) -> Token[Context]:
         """Sets the current `Context` object. Returns a
         token that can be used to reset to the previous `Context`.
 
@@ -41,13 +44,13 @@ class ContextVarsRuntimeContext(_RuntimeContext):
         """Returns the current `Context` object."""
         return self._current_context.get()
 
-    def detach(self, token: object) -> None:
+    def detach(self, token: Token[Context]) -> None:
         """Resets Context to a previous value
 
         Args:
             token: A reference to a previous Context.
         """
-        self._current_context.reset(token)  # type: ignore
+        self._current_context.reset(token)
 
 
 __all__ = ["ContextVarsRuntimeContext"]